+++ src/main/java/com/example/bak/feed/application/command/port/CommunityCommandPort.java
@@ -1,4 +1,4 @@
-package com.example.bak.feed.application.port;
+package com.example.bak.feed.application.command.port;
 
 import com.example.bak.community.domain.Community;
 import java.util.Optional;
