+++ src/main/java/com/example/bak/feed/application/command/port/FeedCommandPort.java
@@ -1,0 +1,5 @@
+package com.example.bak.feed.application.command.port;
+
+public interface FeedCommandPort {
+
+}
