+++ src/main/java/com/example/bak/feed/infra/query/FeedQueryAdapter.java
@@ -1,0 +1,5 @@
+package com.example.bak.feed.infra.query;
+
+public class FeedQueryAdapter {
+
+}
