+++ src/main/java/com/example/bak/feed/presentation/FeedController.java
@@ -1,9 +1,9 @@
 package com.example.bak.feed.presentation;
 
 import com.example.bak.feed.application.FeedService;
-import com.example.bak.feed.application.dto.FeedDetail;
-import com.example.bak.feed.application.dto.FeedResult;
-import com.example.bak.feed.application.dto.FeedSummary;
+import com.example.bak.feed.application.query.dto.FeedDetail;
+import com.example.bak.feed.application.query.dto.FeedResult;
+import com.example.bak.feed.application.query.dto.FeedSummary;
 import com.example.bak.feed.presentation.dto.FeedRequest;
 import com.example.bak.global.common.response.ApiResponse;
 import com.example.bak.global.common.response.ApiResponseFactory;
