+++ src/test/java/com/example/bak/feed/application/FeedServiceUnitTest.java
@@ -6,9 +6,9 @@ import static org.assertj.core.api.Assertions.assertThat;
 import com.example.bak.community.domain.Community;
 import com.example.bak.community.domain.CommunityRepositoryStub;
 import com.example.bak.company.domain.Company;
-import com.example.bak.feed.application.dto.FeedDetail;
-import com.example.bak.feed.application.dto.FeedResult;
-import com.example.bak.feed.application.dto.FeedSummary;
+import com.example.bak.feed.application.query.dto.FeedDetail;
+import com.example.bak.feed.application.query.dto.FeedResult;
+import com.example.bak.feed.application.query.dto.FeedSummary;
 import com.example.bak.feed.domain.Feed;
 import com.example.bak.feed.domain.FeedRepositoryStub;
 import com.example.bak.global.exception.ErrorCode;
